+++ packages/vtable-export/demo/resource-url/image.ts
@@ -1,0 +1,11 @@
+export const animalImageUrl = 'http://' + window.location.host + '/mock-data/animal.jpeg';
+export const animal1ImageUrl = 'http://' + window.location.host + '/mock-data/animal1.jpeg';
+export const animal2ImageUrl = 'http://' + window.location.host + '/mock-data/animal2.jpeg';
+export const animal3ImageUrl = 'http://' + window.location.host + '/mock-data/animal3.jpeg';
+export const animal4ImageUrl = 'http://' + window.location.host + '/mock-data/animal4.jpeg';
+export const flowerImageUrl = 'http://' + window.location.host + '/mock-data/custom-render/flower.jpg';
+export const wolfImageUrl = 'http://' + window.location.host + '/mock-data/custom-render/wolf.jpg';
+export const rabbitImageUrl = 'http://' + window.location.host + '/mock-data/custom-render/rabbit.jpg';
+export const catImageUrl = 'http://' + window.location.host + '/mock-data/custom-render/cat.jpg';
+export const bearImageUrl = 'http://' + window.location.host + '/mock-data/custom-render/bear.jpg';
+export const birdImageUrl = 'http://' + window.location.host + '/mock-data/custom-render/bird.jpg';