+++ packages/vtable/src/core/BaseTable.ts
@@ -2813,7 +2813,7 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
       return false;
     }
     const index = table.getRecordShowIndexByCell(col, row);
-    return table.internalProps.dataSource.hasField(index, field);
+    return table.internalProps.dataSource?.hasField(index, field);
   }
   /**
    * 获取单元格的样式 内部逻辑使用 获取到的样式并不是计算后的
@@ -3610,5 +3610,19 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
     console.error(new ReferenceError(`stage is not defined`));
 
     return null;
+  }
+  /** 根据表格单元格的行列号 获取在body部分的列索引及行索引 */
+  getBodyIndexByTableIndex(col: number, row: number) {
+    return {
+      col: col - this.rowHeaderLevelCount,
+      row: row - this.columnHeaderLevelCount
+    };
+  }
+  /** 根据body部分的列索引及行索引，获取单元格的行列号 */
+  getTableIndexByBodyIndex(col: number, row: number) {
+    return {
+      col: col + this.rowHeaderLevelCount,
+      row: row + this.columnHeaderLevelCount
+    };
   }
 }