+++ packages/vtable/src/scenegraph/component/cell-content.ts
@@ -42,6 +42,7 @@ export class CellContent extends Group {
   _autoWidth = false;
   _autoHeight = false;
   _cellWidth: number;
+  _originCellWidth: number;
   _cellHeight: number;
   _align: CanvasTextAlign;
   _baseline: CanvasTextBaseline;
@@ -77,6 +78,7 @@ export class CellContent extends Group {
     this._autoWidth = option.autoWidth;
     this._autoHeight = option.autoHeight;
     this._cellWidth = option.cellWidth;
+    this._originCellWidth = option.cellWidth;
     this._cellHeight = option.cellHeight;
     this._align = option.align;
     this._baseline = option.baseline;
@@ -233,11 +235,13 @@ export class CellContent extends Group {
 
   updateWidth(width: number) {
     this._cellWidth = width;
+    this._originCellWidth = width;
     this.layout();
   }
 
   updateHeight(height: number) {
     this._cellHeight = height;
+    this._cellWidth = this._originCellWidth;
     this.layout();
   }
 }