+++ packages/vtable/src/scenegraph/group-creater/cell-helper.ts
@@ -616,7 +616,7 @@ function updateCellContent(
     value = table.getCellValue(col, row);
   }
   //解决报错 getCellByCache递归调用 死循环问题
-  if (oldCellGroup.row !== row || oldCellGroup.col !== col) {
+  if (!addNew && (oldCellGroup.row !== row || oldCellGroup.col !== col)) {
     return null;
   }
   const newCellGroup = createCell(